dag-kode/src/uk/dangrew/kode/javafx/time/TimestampBox.java
package uk.dangrew.kode.javafx.time;

import java.time.LocalDateTime;
import java.util.Optional;

import javafx.scene.control.TextField;
import javafx.scene.paint.Color;
import uk.dangrew.kode.datetime.TimestampFormat;
import uk.dangrew.kode.javafx.style.JavaFxStyle;

/**
 * Graphical widget that alows a timestamp to be entered, either spinning or typing, with validation.
 */
public class TimestampBox extends TextField {

   static final Color INVALID_BORDER_COLOUR = Color.RED;
   static final Color VALID_BORDER_COLOUR = Color.GREEN;
   
   private final JavaFxStyle styling;
   private final TimestampFormat format;
   
   public TimestampBox() {
      this( new CaretPositionTextSpinner() );
   }//End Constructor
   
   TimestampBox( CaretPositionTextSpinner spinner ) {
      this.styling = new JavaFxStyle();
      this.format = new TimestampFormat();
      this.setPromptText( format.pattern() );
      this.textProperty().addListener( ( s, o, n ) -> validate() );
      spinner.register( this );
      
      this.validate();
      this.resetInputToNow();
   }//End Constructor
   
   public void resetInputToNow(){
      LocalDateTime now = LocalDateTime.now();
      setText( format.toTimestampString( now ) );
   }//End Method
   
   private void validate(){
      showValidation( isValid() );
   }//End Method
   
   boolean isValid(){
      return Optional.ofNullable( parseText() ).isPresent();
   }//End Method
   
   private LocalDateTime parseText() {
      return format.parseTimestamp( getText() );
   }//End Method
   
   private void showValidation( boolean valid ) {
      if ( valid ) {
         setBorder( styling.borderFor( VALID_BORDER_COLOUR ) );
      } else {
         setBorder( styling.borderFor( INVALID_BORDER_COLOUR ) );
      }
   }//End Method
   
   public LocalDateTime getTextValue(){
      return parseText();
   }//End Method
   
}//End Class
